+++ pkg/metadata/save.go
@@ -11,27 +11,15 @@ func (tm *TableMetadata) Save(location string, metadataOnly bool) (uint64, error
 	newTM := TableMetadata{
 		Table:                tm.Table,
 		Database:             tm.Database,
-		IncrementOf:          tm.IncrementOf,
 		Query:                tm.Query,
 		DependenciesTable:    tm.DependenciesTable,
 		DependenciesDatabase: tm.DependenciesDatabase,
 		MetadataOnly:         true,
 	}
-	parts := map[string][]Part{}
-	for disk, p := range tm.Parts {
-		newp := make([]Part, len(p))
-		for i := range p {
-			newp[i] = Part{
-				Name:     p[i].Name,
-				Required: p[i].Required,
-			}
-		}
-		parts[disk] = newp
-	}
 
 	if !metadataOnly {
 		newTM.Files = tm.Files
-		newTM.Parts = parts
+		newTM.Parts = tm.Parts
 		newTM.Size = tm.Size
 		newTM.TotalBytes = tm.TotalBytes
 		newTM.MetadataOnly = false